core: move message notification to home handler
