+++ packages/hydrooj/src/plugin-api.ts
@@ -6,6 +6,7 @@ export { default as moment, isMoment } from 'moment-timezone';
 
 export {
     WebService, Router, HandlerCommon, httpServer,
+    Query, Mutation, Subscription, Apis,
 } from '@hydrooj/framework';
 
 export * from './pipelineUtils';
